feat: 改造core结构 1. 插件化存储核心 2. 引入viper作为配置文件代理，固定配置迁移至配置文件
