--- core/main.go
@@ -1,14 +1,0 @@
-package main
-
-import (
-	"github.com/Trinoooo/eggie_kv/core/cli"
-	"log"
-	"os"
-)
-
-func main() {
-	wrapper := cli.NewWrapper()
-	if err := wrapper.Run(os.Args); err != nil {
-		log.Fatalln(err)
-	}
-}
