+++ kv_storage/main.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+	"github.com/Trinoooo/eggie_kv/kv_storage/cli"
+	"log"
+	"os"
+)
+
+func main() {
+	wrapper := cli.NewWrapper()
+	if err := wrapper.Run(os.Args); err != nil {
+		log.Fatalln(err)
+	}
+}
