+++ makefile
@@ -1,13 +1,13 @@
 ProjectPath=$(GOPATH)/src/github.com/Trinoooo/eggieKv
 
 # 构建产物
-build-core:
-	cd $(ProjectPath)/core && \
+build-kv-storage:
+	cd $(ProjectPath)/kv_storage && \
 	go mod tidy && \
 	go build -o eggie_kv_server main.go
 
 build-interactive-cli:
-	cd $(ProjectPath)/interactive/cli && \
+	cd $(ProjectPath)/cli && \
 	go mod tidy && \
 	go build -o eggie_kv_client main.go
